Collapse similar IPFS utility files
