+++ browser/net/ipfs_redirect_network_delegate_helper_unittest.cc
@@ -15,7 +15,6 @@
 #include "brave/common/translate_network_constants.h"
 #include "brave/components/ipfs/features.h"
 #include "brave/components/ipfs/ipfs_constants.h"
-#include "brave/components/ipfs/ipfs_gateway.h"
 #include "brave/components/ipfs/ipfs_utils.h"
 #include "brave/components/ipfs/pref_names.h"
 #include "chrome/common/channel_info.h"
