+++ components/ipfs/ipfs_utils_unittest.cc
@@ -9,7 +9,6 @@
 #include <vector>
 
 #include "brave/components/ipfs/ipfs_constants.h"
-#include "brave/components/ipfs/ipfs_gateway.h"
 #include "brave/components/ipfs/pref_names.h"
 #include "chrome/common/channel_info.h"
 #include "components/prefs/pref_registry_simple.h"
@@ -32,7 +31,13 @@ class IpfsUtilsUnitTest : public testing::Test {
         kIPFSResolveMethod,
         static_cast<int>(ipfs::IPFSResolveMethodTypes::IPFS_ASK));
     user_prefs::UserPrefs::Set(browser_context_.get(), &prefs_);
+    public_gateway_ = GURL(ipfs::kDefaultIPFSGateway);
+    local_gateway_ = GURL("http://localhost:48080");
   }
+
+  const GURL& local_gateway() { return local_gateway_; }
+  const GURL& public_gateway() { return public_gateway_; }
+  PrefService* prefs() { return &prefs_; }
 
   void SetIPFSResolveMethodPref(ipfs::IPFSResolveMethodTypes type) {
     prefs_.SetInteger(kIPFSResolveMethod, static_cast<int>(type));
@@ -44,6 +49,8 @@ class IpfsUtilsUnitTest : public testing::Test {
   content::BrowserTaskEnvironment task_environment_;
   std::unique_ptr<content::TestBrowserContext> browser_context_;
   TestingPrefServiceSimple prefs_;
+  GURL local_gateway_;
+  GURL public_gateway_;
 };
 
 TEST_F(IpfsUtilsUnitTest, HasIPFSPath) {
@@ -204,4 +211,266 @@ TEST_F(IpfsUtilsUnitTest, ResolveIPFSURI) {
                                    &gateway_url));
   ASSERT_EQ(gateway_url, GURL("http://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgp"
                               "dr4cjr3oz3evfyavhwq.ipfs.localhost:48080"));
+}
+
+TEST_F(IpfsUtilsUnitTest, GetDefaultIPFSGateway) {
+  prefs()->SetString(kIPFSPublicGatewayAddress, "https://example.com/");
+  EXPECT_EQ(ipfs::GetDefaultIPFSGateway(context()),
+            GURL("https://example.com/"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURINotIPFSScheme) {
+  GURL url("http://a.com/ipfs/QmfM2r8seH2GiRaC4esTjeraXEachRt8ZsSeGaWTPLyMoG");
+  GURL new_url;
+  ASSERT_FALSE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSScheme) {
+  GURL url("ipfs://QmfM2r8seH2GiRaC4esTjeraXEachRt8ZsSeGaWTPLyMoG");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url, GURL("https://dweb.link/ipfs/"
+                          "QmfM2r8seH2GiRaC4esTjeraXEachRt8ZsSeGaWTPLyMoG"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPNSScheme) {
+  GURL url("ipns://QmSrPmbaUKA3ZodhzPWZnpFgcPMFWF4QsxXbkWfEptTBJd");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url, GURL("https://dweb.link/ipns/"
+                          "QmSrPmbaUKA3ZodhzPWZnpFgcPMFWF4QsxXbkWfEptTBJd"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeLocal) {
+  GURL url("ipfs://QmfM2r8seH2GiRaC4esTjeraXEachRt8ZsSeGaWTPLyMoG");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), false));
+  EXPECT_EQ(new_url, GURL("http://localhost:48080/ipfs/"
+                          "QmfM2r8seH2GiRaC4esTjeraXEachRt8ZsSeGaWTPLyMoG"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPNSSchemeLocal) {
+  GURL url("ipns://QmSrPmbaUKA3ZodhzPWZnpFgcPMFWF4QsxXbkWfEptTBJd");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), false));
+  EXPECT_EQ(new_url, GURL("http://localhost:48080/ipns/"
+                          "QmSrPmbaUKA3ZodhzPWZnpFgcPMFWF4QsxXbkWfEptTBJd"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPath) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("https://dweb.link/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathAndHash) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html#Emerging_artist");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("https://dweb.link/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html#Emerging_artist"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeLocalWithPathAndHash) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html#Emerging_artist");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("http://localhost:48080/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html#Emerging_artist"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathAndQuery) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("https://dweb.link/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html?test=true"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeLocalWithPathAndQuery) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("http://localhost:48080/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html?test=true"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathQueryHash) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true#test");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("https://dweb.link/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html?test=true#test"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeLocalWithPathQueryHash) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true#test");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), false));
+  EXPECT_EQ(new_url,
+            GURL("http://localhost:48080/ipfs/"
+                 "bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+                 "/wiki/Vincent_van_Gogh.html?test=true#test"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURINotIPFSSchemeSubdomain) {
+  GURL url(
+      "http://a.com/ipfs/bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsgg"
+      "enkbw6slwk4");
+  GURL new_url;
+  ASSERT_FALSE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeSubdomain) {
+  GURL url(
+      "ipfs://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw6slwk4");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("https://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkb"
+                 "w6slwk4.ipfs.dweb.link/"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPNSSchemeSubdomain) {
+  GURL url(
+      "ipns://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw6slwk4");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("https://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkb"
+                 "w6slwk4.ipns.dweb.link/"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeLocalSubdomain) {
+  GURL url(
+      "ipfs://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw6slwk4");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("http://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw"
+                 "6slwk4.ipfs.localhost:48080/"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPNSSchemeLocalSubdomain) {
+  GURL url(
+      "ipns://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw6slwk4");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), true));
+  EXPECT_EQ(new_url, GURL("http://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanoko"
+                          "nsggenkbw6slwk4.ipns.localhost:48080/"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathSubdomain) {
+  GURL url(
+      "ipfs://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkbw6slwk4"
+      "/wiki/Vincent_van_Gogh.html");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("https://bafybeiffndsajwhk3lwjewwdxqntmjm4b5wxaaanokonsggenkb"
+                 "w6slwk4.ipfs.dweb.link/wiki/Vincent_van_Gogh.html"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathAndHashSubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html#Emerging_artist");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(
+      new_url,
+      GURL(
+          "https://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3ev"
+          "fyavhwq.ipfs.dweb.link/wiki/Vincent_van_Gogh.html#Emerging_artist"));
+}
+
+TEST_F(IpfsUtilsUnitTest,
+       TranslateIPFSURIIPFSSchemeLocalWithPathAndHashSubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html#Emerging_artist");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("http://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evf"
+                 "yavhwq.ipfs.localhost:48080/wiki/Vincent_van_Gogh.html"
+                 "#Emerging_artist"));
+}
+
+TEST_F(IpfsUtilsUnitTest, TranslateIPFSURIIPFSSchemeWithPathAndQuerySubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("https://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evf"
+                 "yavhwq.ipfs.dweb.link/wiki/Vincent_van_Gogh.html?test=true"));
+}
+
+TEST_F(IpfsUtilsUnitTest,
+       TranslateIPFSURIIPFSSchemeLocalWithPathAndQuerySubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("http://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evf"
+                 "yavhwq.ipfs.localhost:48080/wiki/Vincent_van_Gogh.html"
+                 "?test=true"));
+}
+
+TEST_F(IpfsUtilsUnitTest,
+       TranslateIPFSURIIPFSSchemeWithPathQueryHashSubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true#test");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, public_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("https://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3ev"
+                 "fyavhwq.ipfs.dweb.link/wiki/Vincent_van_Gogh.html?test=true"
+                 "#test"));
+}
+
+TEST_F(IpfsUtilsUnitTest,
+       TranslateIPFSURIIPFSSchemeLocalWithPathQueryHashSubdomain) {
+  GURL url(
+      "ipfs://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evfyavhwq"
+      "/wiki/Vincent_van_Gogh.html?test=true#test");
+  GURL new_url;
+  ASSERT_TRUE(ipfs::TranslateIPFSURI(url, &new_url, local_gateway(), true));
+  EXPECT_EQ(new_url,
+            GURL("http://bafybeiemxf5abjwjbikoz4mc3a3dla6ual3jsgpdr4cjr3oz3evf"
+                 "yavhwq.ipfs.localhost:48080/wiki/Vincent_van_Gogh.html"
+                 "?test=true#test"));
 }
